refactor: align playstation auth method naming
