+++ turbo/packages/connectors/src/auth-providers/connectors/playstation/api.ts
